Add spectest-interp tool; split from wasm-interp
* `wasm-interp` tool now only runs `.wasm` file, not spec tests
* `wasm-interp` has a new flag `--host-print` for importing a print
  function named "host.print"
* `spectest-interp` tool runs only `.json` files
