+++ test/find_exe.py
@@ -25,7 +25,7 @@ SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
 REPO_ROOT_DIR = os.path.dirname(SCRIPT_DIR)
 EXECUTABLES = [
     'wat2wasm', 'wast2json', 'wasm2wat', 'wasm-objdump', 'wasm-interp',
-    'wasm-opcodecnt', 'wat-desugar', 'wasm-link'
+    'wasm-opcodecnt', 'wat-desugar', 'wasm-link', 'spectest-interp',
 ]
 
 
@@ -86,6 +86,10 @@ def GetWasmlinkExecutable(override=None):
 
 def GetWasmInterpExecutable(override=None):
   return FindExecutable('wasm-interp', override)
+
+
+def GetSpectestInterpExecutable(override=None):
+  return FindExecutable('spectest-interp', override)
 
 
 def GetWasmOpcodeCntExecutable(override=None):
