+++ test/help/spectest-interp.txt
@@ -1,0 +1,21 @@
+;;; EXE: %(spectest-interp)s
+;;; FLAGS: --help
+(;; STDOUT ;;;
+usage: spectest-interp [options] filename
+
+  read a Spectest JSON file, and run its tests in the interpreter.
+
+examples:
+  # parse test.json and run the spec tests
+  $ spectest-interp test.json
+
+options:
+  -v, --verbose                               Use multiple times for more info
+  -h, --help                                  Print this help message
+      --enable-exceptions                     Experimental exception handling
+      --enable-saturating-float-to-int        Saturating float-to-int operators
+      --enable-threads                        Threading support
+  -V, --value-stack-size=SIZE                 Size in elements of the value stack
+  -C, --call-stack-size=SIZE                  Size in elements of the call stack
+  -t, --trace                                 Trace execution
+;;; STDOUT ;;)
