+++ test/interp/callimport-zero-args.txt
@@ -1,12 +1,13 @@
 ;;; TOOL: run-interp
+;;; FLAGS: --host-print
 (module
-  (import "spectest" "print" (func $imported (result i32)))
+  (import "host" "print" (func $imported (result i32)))
 
   (func (export "f") (result i32)
     i32.const 13
     call $imported
     i32.add))
 (;; STDOUT ;;;
-called host spectest.print() => i32:0
+called host host.print() => i32:0
 f() => i32:13
 ;;; STDOUT ;;)
