+++ test/interp/import.txt
@@ -1,7 +1,8 @@
 ;;; TOOL: run-interp
+;;; FLAGS: --host-print
 (module
-  (import "spectest" "print" (func $print_i32 (param i32)))
-  (import "spectest" "print" (func $print_i32_i32 (param i32 i32)))
+  (import "host" "print" (func $print_i32 (param i32)))
+  (import "host" "print" (func $print_i32_i32 (param i32 i32)))
   (func (export "test") (result i32)
     i32.const 100
     call $print_i32
@@ -12,7 +13,7 @@
     return)
 )
 (;; STDOUT ;;;
-called host spectest.print(i32:100) =>
-called host spectest.print(i32:200, i32:300) =>
+called host host.print(i32:100) =>
+called host host.print(i32:200, i32:300) =>
 test() => i32:1
 ;;; STDOUT ;;)
